Clean up contract utils
 - Removes two `utils` files as coding contract structure now allows for each contract to be more self-contained. Note BigInt helper is retained as it may be used outside of its associated contract at some point.
 - Continues the Sisyphean attempts at clarification of the contract wording.
